{"schema":"decodex/commit/1","summary":"Extract frozen annotation session state owner","authority":"manual"}
